# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/languages": {
            "get": {
                "tags": [
                    "Languages"
                ],
                "summary": "Get languages list",
                "operationId": "13f72dfeebea6adfa84684cf31bade15",
                "responses": {
                    "200": {
                        "description": "language response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Language"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/languages/{lang}": {
            "get": {
                "tags": [
                    "Languages"
                ],
                "summary": "Get language by lang",
                "operationId": "102bf94e2acc5ac55f646cec845c9101",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "laтg to fetch",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "language response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Language"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lrs": {
            "get": {
                "tags": [
                    "Lr"
                ],
                "summary": "Get lrs list",
                "operationId": "9bc225375276dc1ad2771727be9ce9a8",
                "responses": {
                    "200": {
                        "description": "lr response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Location"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lrs/{lr}": {
            "get": {
                "tags": [
                    "Lr"
                ],
                "summary": "Get location by lr",
                "operationId": "71be1c166f1fca4cbaee03d9a44df6cc",
                "parameters": [
                    {
                        "name": "lr",
                        "in": "path",
                        "description": "lr to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lr response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lrs/search/{search}": {
            "get": {
                "tags": [
                    "Lr"
                ],
                "summary": "Get lr search",
                "operationId": "b7e93af3595058a287e9586af188ce3a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "search",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Мос"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lr response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Language": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "lang": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "aparser": {
                                "type": "string"
                            },
                            "three": {
                                "type": "string"
                            },
                            "rucaptcha": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Location": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "lr": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "parent_id": {
                                "type": "string"
                            },
                            "parent_string": {
                                "type": "string"
                            },
                            "sort": {
                                "type": "integer"
                            },
                            "pop": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}